pyproject.toml
[project]
name = "raindrop_digest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "httpx>=0.27.2",
    "openai>=1.57.0",
    "pytest>=8.4.2",
    "readability-lxml>=0.8.1",
    "sendgrid>=6.11.0",
]
